+++ merge/merge.go
@@ -1,11 +1,16 @@
 package merge
 
 import (
+	"encoding/json"
 	"fmt"
+	"io"
 	"log"
 	"os"
 	"os/exec"
 	"path/filepath"
+	"regexp"
+	"strconv"
+	"strings"
 
 	arg "github.com/alexflint/go-arg"
 	"github.com/brentp/smoove/shared"
@@ -17,6 +22,127 @@ type cliargs struct {
 	OutDir string   `arg:"-o,help:output directory."`
 	Fasta  string   `arg:"-f,required,help:fasta file."`
 	VCFs   []string `arg:"positional,required,help:path to vcfs."`
+}
+
+type count struct {
+	sample       string
+	split_before int
+	disc_before  int
+	split_after  int
+	disc_after   int
+}
+
+func mustInt(s string) int {
+	r, err := strconv.Atoi(s)
+	if err != nil {
+		panic(err)
+	}
+	return r
+
+}
+
+const tmpl = `
+<!DOCTYPE html>
+<html lang="en">
+<head>
+<script type="text/javascript" src="https://code.jquery.com/jquery-3.3.1.min.js"></script>
+<script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
+</head>
+<body>
+<div id="before"></div>
+<div id="after"></div>
+<script>
+
+var tracebefore = {
+  x: %s,
+  y: %s,
+  text: %s,
+  mode: 'markers',
+  marker: {
+    size: 10,
+    color: %s,
+  }
+};
+var traceafter = {
+  x: %s,
+  y: %s,
+  text: %s,
+  mode: 'markers',
+  marker: {
+    size: 10,
+    color: %s,
+  }
+};
+
+var layout = {title: 'number of reads before filtering', hovermode: 'closest'}
+layout.xaxis = {title : "split reads"}
+layout.yaxis = {title : "discordant reads"}
+
+Plotly.newPlot('before', [tracebefore], layout);
+layout.title = 'number of reads after filtering';
+Plotly.newPlot('after', [traceafter], layout);
+</script>
+Created by <a href="https://github.com/brentp/smoove">smoove</a>
+</body>
+</html>
+`
+
+func mustMarshal(x interface{}) string {
+	v, err := json.Marshal(x)
+	if err != nil {
+		panic(err)
+	}
+	return string(v)
+}
+
+func plotCounts(path string, outpath string) {
+	// see: https://plot.ly/javascript/line-and-scatter/
+	f, err := xopen.Ropen(path)
+	defer f.Close()
+	if err != nil {
+		panic(err)
+	}
+
+	split_before := make([]int, 0, 16)
+	split_after := make([]int, 0, 16)
+	disc_before := make([]int, 0, 16)
+	disc_after := make([]int, 0, 16)
+	samples := make([]string, 0, 16)
+	patt := regexp.MustCompile("[,:]")
+
+	for {
+		line, err := f.ReadString('\n')
+		if err == io.EOF {
+			break
+		}
+		if line[0] != '#' {
+			break
+		}
+		if !strings.HasPrefix(line, "##smoove_count_stats=") {
+			continue
+		}
+		// ##smoove_count_stats=sample:4494396,7960884,747932,297854
+		tmp := strings.Split(line, "=")
+		info := patt.Split(strings.TrimRight(tmp[1], "\r\n"), -1)
+		samples = append(samples, info[0])
+		split_before = append(split_before, mustInt(info[1]))
+		disc_before = append(disc_before, mustInt(info[2]))
+		split_after = append(split_after, mustInt(info[3]))
+		disc_after = append(disc_after, mustInt(info[4]))
+	}
+	rng := make([]int, len(disc_after))
+	for i := 0; i < len(rng); i++ {
+		rng[i] = i
+	}
+	wtr, err := xopen.Wopen(outpath)
+	if err != nil {
+		panic(err)
+	}
+
+	fmt.Fprintf(wtr, tmpl, mustMarshal(split_before), mustMarshal(disc_before), mustMarshal(samples), mustMarshal(rng), mustMarshal(split_after), mustMarshal(disc_after), mustMarshal(samples), mustMarshal(rng))
+	wtr.Close()
+	shared.Slogger.Printf("wrote html file of disc, split counts to %s", wtr.Name())
+
 }
 
 func Main() {
@@ -51,4 +177,5 @@ func Main() {
 	}
 	os.Remove(f.Name())
 	shared.Slogger.Printf("wrote sites file to %s", of)
+	plotCounts(of, filepath.Join(cli.OutDir, cli.Name)+".smoove-counts.html")
 }